refactor: Create general test for common features
Create `test_general` to test common features used across all the cli.
Such as:
- Authentication (test if raise error if API returns 401/403)
- SSL validation (ensure error is raised if invalid server cert)
- Timeouts

Since all the subcommands use the same api backend, this test_general
use `dig` as a vehicle to test all the common features.
